refactor(autotune): разбить монолитный модуль на подмодули и добавить кэширование DNS
